NavMenu wip (almost done)
