+++ examples/Demo/Shared/Pages/Lab/IssueTester.razor
@@ -1,31 +1,2 @@
 ﻿@page "/IssueTester"
 
-<div style="background: var(--neutral-layer-3); display: flex; padding: 10px;">
-    <FluentNavMenu2 @bind-Expanded="@Expanded" Width="250" Collapsible="true" Title="Collapsible demo">
-        <FluentNavLink Icon="@(new Icons.Regular.Size20.Home())" Href="/NavMenu">Item 1</FluentNavLink>
-        <FluentNavLink Href="/Lab/Overview">Item 2</FluentNavLink>
-        <FluentNavGroup Title="Item 3" Icon="@(new Icons.Regular.Size20.EarthLeaf())" Href="/NavMenu">
-            <FluentNavLink Icon="@(new Icons.Regular.Size20.LeafOne())" Href="/NavMenu">Item 3.1</FluentNavLink>
-            <FluentNavLink Icon="@(new Icons.Regular.Size20.LeafTwo())" Href="/NavMenu">Item 3.2</FluentNavLink>
-        </FluentNavGroup>
-        <FluentNavLink Icon="@(new Icons.Regular.Size20.CalendarAgenda())" Disabled="true" Href="https://microsoft.com">Item 4</FluentNavLink>
-        <FluentNavLink Icon="@(new Icons.Regular.Size20.Home())" Href="/NavMenu" Disabled="true">Item 5</FluentNavLink>
-        <FluentNavGroup Expanded="true" Title="Item 6" Icon="@(new Icons.Regular.Size20.EarthLeaf())" >
-            <FluentNavLink Icon="@(new Icons.Regular.Size20.LeafOne())" Href="/NavMenu">Item 6.1</FluentNavLink>
-            <FluentNavLink Icon="@(new Icons.Regular.Size20.LeafTwo())" Href="/NavMenu">Item 6.2</FluentNavLink>
-            <FluentNavGroup Expanded="true" Title="Item 6.3" Icon="@(new Icons.Regular.Size20.EarthLeaf())">
-                <FluentNavLink Icon="@(new Icons.Regular.Size20.LeafOne())" Href="/NavMenu">Item 6.3.1</FluentNavLink>
-                <FluentNavLink Icon="@(new Icons.Regular.Size20.LeafTwo())" Href="/NavMenu">Item 6.3.2</FluentNavLink>
-                <FluentNavGroup Expanded="true" Title="Item 6.3.3" Icon="@(new Icons.Regular.Size20.EarthLeaf())">
-                    <FluentNavLink Icon="@(new Icons.Regular.Size20.LeafOne())" Href="/NavMenu">Item 6.3.3.1</FluentNavLink>
-                    <FluentNavLink Icon="@(new Icons.Regular.Size20.LeafTwo())" Href="/NavMenu">Item 6.3.3.2</FluentNavLink>
-                </FluentNavGroup>
-            </FluentNavGroup>
-        </FluentNavGroup>
-    </FluentNavMenu2>
-</div>
-
-@code
-{
-    bool Expanded = true;
-}
