+++ examples/Demo/Shared/Pages/NavMenu/Examples/NavMenuDataBound.razor
@@ -1,20 +1,16 @@
-﻿@namespace FluentUI.Demo.Shared
-
-@inject ILogger<NavMenuDefault> logger;
-
-<h2>Navigation Examples</h2>
+﻿<h2>Navigation Examples</h2>
 
 <FluentStack Orientation="Orientation.Horizontal">
     <!-- Menu with sub-items and icons -->
     <FluentNavMenu @bind-Expanded=MenuExpanded Title="Databound demo">
-        <FluentNavMenuGroup Id="Group1" Text="Item 1" @bind-Expanded=Item1Expanded Icon="@(new Icons.Regular.Size24.LeafOne())" Href="/NavMenu">
-            <FluentNavMenuLink Text="Item 1.1" Icon="@(new Icons.Regular.Size24.LeafOne())" @bind-Selected=Item1Point1Selected Href="/NavMenu" />
-            <FluentNavMenuLink Text="Item 1.2" Icon="@(new Icons.Regular.Size24.LeafTwo())" @bind-Selected=Item1Point2Selected Href="/NavMenu" />
-        </FluentNavMenuGroup>
-        <FluentNavMenuGroup Id="Group2" Text="Item 2" @bind-Expanded=Item2Expanded Icon="@(new Icons.Regular.Size24.LeafTwo())" Href="/NavMenu">
-            <FluentNavMenuLink Text="Item 2.1" Icon="@(new Icons.Regular.Size24.LeafOne())" @bind-Selected=Item2Point1Selected Href="/NavMenu" />
-            <FluentNavMenuLink Text="Item 2.2" Icon="@(new Icons.Regular.Size24.LeafTwo())" @bind-Selected=Item2Point2Selected Href="/NavMenu" />
-        </FluentNavMenuGroup>
+        <FluentNavGroup Id="Group1" Title="Item 1" @bind-Expanded=Item1Expanded Icon="@(new Icons.Regular.Size24.LeafOne())">
+            <FluentNavLink Icon="@(new Icons.Regular.Size24.LeafOne())">Item 1.1</FluentNavLink>
+            <FluentNavLink Icon="@(new Icons.Regular.Size24.LeafTwo())">Item 1.2</FluentNavLink>
+        </FluentNavGroup>
+        <FluentNavGroup Id="Group2" Title="Item 2" @bind-Expanded=Item2Expanded Icon="@(new Icons.Regular.Size24.LeafTwo())">
+            <FluentNavLink Icon="@(new Icons.Regular.Size24.LeafOne())">Item 2.1</FluentNavLink>
+            <FluentNavLink Icon="@(new Icons.Regular.Size24.LeafTwo())">Item 2.2</FluentNavLink>
+        </FluentNavGroup>
     </FluentNavMenu>
     <FluentStack Orientation="Orientation.Vertical">
 
@@ -29,19 +25,7 @@
             <span aria-label="Item 2 expanded">Item 2</span>
         </FluentCheckbox>
 
-        <h2>Selected elements</h2>
-        <FluentCheckbox @bind-Value=Item1Point1Selected>
-            <span aria-label="Item 1.1 selected">Item 1.1</span>
-        </FluentCheckbox>
-        <FluentCheckbox @bind-Value=Item1Point2Selected>
-            <span aria-label="Item 1.2 selected">Item 1.2</span>
-        </FluentCheckbox>
-        <FluentCheckbox @bind-Value=Item2Point1Selected>
-            <span aria-label="Item 2.1 selected">Item 2.1</span>
-        </FluentCheckbox>
-        <FluentCheckbox @bind-Value=Item2Point2Selected>
-            <span aria-label="Item 2.2 selected">Item 2.2</span>
-        </FluentCheckbox>
+        
 
     </FluentStack>
 </FluentStack>
@@ -50,11 +34,6 @@
     bool MenuExpanded = true;
     bool Item1Expanded = true;
     bool Item2Expanded = true;
-
-    bool Item1Point1Selected = false;
-    bool Item1Point2Selected = false;
-    bool Item2Point1Selected = false;
-    bool Item2Point2Selected = false;
 }
