+++ examples/Demo/Shared/Pages/NavMenu/Examples/NavMenuDefault.razor
@@ -1,71 +1,62 @@
-﻿@namespace FluentUI.Demo.Shared
-
-@inject ILogger<NavMenuDefault> logger;
-
-<h2>Navigation Examples</h2>
+﻿<h2>Navigation Examples</h2>
 
 <FluentStack Orientation="Orientation.Horizontal" >
-    <!-- Menu with sub-items and icons -->
-    <FluentNavMenu Collapsible="false" Title="Default demo">
-        <FluentNavMenuLink Text="Item 1" Icon="@(new Icons.Regular.Size24.Home())" Href="/NavMenu" />
-        <FluentNavMenuLink Text="Item 2" Href="/NavMenu" />
-        <FluentNavMenuGroup Text="Item 3" Href="/NavMenu">
-            <FluentNavMenuLink Text="Item 3.1" Icon="@(new Icons.Regular.Size24.LeafOne())" Href="/NavMenu" />
-            <FluentNavMenuLink Text="Item 3.2" Icon="@(new Icons.Regular.Size24.LeafTwo())" Href="/NavMenu" />
-        </FluentNavMenuGroup>
-        <FluentNavMenuGroup Text="Item 4" @onclick="OnClick" Href="/NavMenu">
-            <FluentNavMenuLink Text="Item 4.1" @onclick="OnClick" Href="/NavMenu" />
-            <FluentNavMenuLink Text="Item 4.2" @onclick="OnClick" Icon="@(new Icons.Regular.Size24.LeafOne())" Href="/NavMenu" />
-        </FluentNavMenuGroup>
-        <FluentNavMenuGroup Text="Item 5" @onclick="OnClick" Href="/NavMenu">
-            <FluentNavMenuLink Text="Item 5.1" @onclick="OnClick" Icon="@(new Icons.Regular.Size24.LeafTwo())" Href="/NavMenu" />
-            <FluentNavMenuGroup @onclick="OnClick" Text="Item 5.2" Href="/NavMenu">
-                <FluentNavMenuLink Text="Item 5.2.1" @onclick="OnClick" Href="/NavMenu" />
-                <FluentNavMenuLink Text="Item 5.2.2" @onclick="OnClick" Href="/NavMenu" />
-            </FluentNavMenuGroup>
-            <FluentNavMenuGroup @onclick="OnClick" Text="Item 5.3" Href="/NavMenu">
-                <FluentNavMenuLink Text="Item 5.3.1" @onclick="OnClick" Icon="@(new Icons.Regular.Size24.LeafOne())" Href="/NavMenu" />
-                <FluentNavMenuLink Text="Item 5.3.2" @onclick="OnClick" Href="/NavMenu" />
-            </FluentNavMenuGroup>
-        </FluentNavMenuGroup>
-        <FluentNavMenuGroup @onclick="OnClick" Text="Items with Href">
-            <FluentNavMenuLink Text="Go to Bing" @onclick="OnClick" Icon="@(new Icons.Regular.Size24.LeafOne())" Href="https://bing.com" />
-            <FluentNavMenuLink Text="Go to home page" @onclick="OnClick" Icon="@(new Icons.Regular.Size24.Home())" Href="/" />
-        </FluentNavMenuGroup>
-        <FluentNavMenuLink Text="Last item" @onclick="OnClick" Icon="@(new Icons.Regular.Size24.LeafThree())" Disabled="true" />
-    </FluentNavMenu>
+    <div style="background: var(--neutral-layer-3); display: flex; padding: 10px;">
+        <FluentNavMenu @bind-Expanded="@expanded" Width="250" Title="Navigation menu">
+            <FluentNavLink Icon="@(new Icons.Regular.Size20.Home())" Href="/" Match="NavLinkMatch.All">Home</FluentNavLink>
+            <FluentNavLink Href="/NavMenu">Item 2</FluentNavLink>
+            <FluentNavGroup OnClick="OnClick" Title="Item 3" Icon="@(new Icons.Regular.Size20.EarthLeaf())">
+                <FluentNavLink OnClick="OnClick" Icon="@(new Icons.Regular.Size20.LeafOne())">Item 3.1</FluentNavLink>
+                <FluentNavLink OnClick="OnClick" Icon="@(new Icons.Regular.Size20.LeafTwo())">Item 3.2</FluentNavLink>
+            </FluentNavGroup>
+            <FluentNavLink Icon="@(new Icons.Regular.Size20.CalendarAgenda())" Disabled="true" Href="https://microsoft.com">Item 4</FluentNavLink>
+            <FluentNavLink Icon="@(new Icons.Regular.Size20.Home())" Disabled="true">Item 5</FluentNavLink>
+            <FluentNavGroup Expanded="true" Title="Item 6" Icon="@(new Icons.Regular.Size20.EarthLeaf())">
+                <FluentNavLink Icon="@(new Icons.Regular.Size20.LeafOne())">Item 6.1</FluentNavLink>
+                <FluentNavLink Icon="@(new Icons.Regular.Size20.LeafTwo())">Item 6.2</FluentNavLink>
+                <FluentNavGroup Expanded="true" Title="Item 6.3" Icon="@(new Icons.Regular.Size20.EarthLeaf())">
+                    <FluentNavLink Icon="@(new Icons.Regular.Size20.LeafOne())">Item 6.3.1</FluentNavLink>
+                    <FluentNavLink Icon="@(new Icons.Regular.Size20.LeafTwo())">Item 6.3.2</FluentNavLink>
+                    <FluentNavGroup Expanded="true" Title="Item 6.3.3" Icon="@(new Icons.Regular.Size20.EarthLeaf())">
+                        <FluentNavLink Icon="@(new Icons.Regular.Size20.LeafOne())">Item 6.3.3.1</FluentNavLink>
+                        <FluentNavLink Disabled="true" Icon="@(new Icons.Regular.Size20.LeafTwo())">Item 6.3.3.2</FluentNavLink>
+                        <FluentNavGroup Disabled="true" Expanded="true" Title="Item 6.3.3.3" Icon="@(new Icons.Regular.Size20.EarthLeaf())">
+                            <FluentNavLink Icon="@(new Icons.Regular.Size20.LeafOne())">Item 6.3.3.3.1</FluentNavLink>
+                            <FluentNavLink Icon="@(new Icons.Regular.Size20.LeafTwo())">Item 6.3.3.3.2</FluentNavLink>
+                        </FluentNavGroup>
+                    </FluentNavGroup>
+                </FluentNavGroup>
+            </FluentNavGroup>
+        </FluentNavMenu>
+    </div>
+
+    @code
+    {
+        bool expanded = true;
+    }
 
     <!-- Menu with icons -->
     <FluentNavMenu>
-        <FluentNavMenuLink Text="Item 1" @onclick="OnClick" Icon="@(new Icons.Regular.Size24.Home())" Href="/NavMenu" />
-        <FluentNavMenuLink Text="Item 2" @onclick="OnClick" Href="/NavMenu" />
-        <FluentNavMenuLink Text="Item 3" @onclick="OnClick" Icon="@(new Icons.Regular.Size24.LeafOne())" Href="/NavMenu" />
-        <FluentNavMenuLink Text="Item 4" @onclick="OnClick" Href="/NavMenu" />
+        <FluentNavLink OnClick="OnClick" Icon="@(new Icons.Regular.Size24.Home())" >Item 1</FluentNavLink>
+        <FluentNavLink OnClick="OnClick" >Item 2</FluentNavLink>
+        <FluentNavLink OnClick="OnClick" Icon="@(new Icons.Regular.Size24.LeafOne())" >Item 3</FluentNavLink>
+        <FluentNavLink OnClick="OnClick" >Item 4</FluentNavLink>
     </FluentNavMenu>
 
     <!-- Menu simple -->
     <FluentNavMenu>
-        <FluentNavMenuLink Text="Item 1" @onclick="OnClick" Href="/NavMenu" />
-        <FluentNavMenuLink Text="Item 2" @onclick="OnClick" Href="/NavMenu" />
-        <FluentNavMenuLink Text="Item 3" @onclick="OnClick" Href="/NavMenu" />
-        <FluentNavMenuLink Text="Item 4" @onclick="OnClick" Href="/NavMenu" />
+        <FluentNavLink OnClick="OnClick" >Item 1</FluentNavLink>
+        <FluentNavLink OnClick="OnClick" >Item 2</FluentNavLink>
+        <FluentNavLink OnClick="OnClick" >Item 3</FluentNavLink>
+        <FluentNavLink OnClick="OnClick" >Item 4</FluentNavLink> 
     </FluentNavMenu>
 
 </FluentStack>
 @code
 {
-
     void OnClick(MouseEventArgs e)
     {
-        logger.LogInformation("Item Clicked");
-    }
-
-    protected override void OnAfterRender(bool firstRender) {
-        if (firstRender) 
-        {
-            
-        }
-        base.OnAfterRender(firstRender);
+        DemoLogger.WriteLine("NavMenu item clicked");
     }
 }
