+++ examples/Demo/Shared/Pages/NavMenuTree/NavMenuTreePage.razor
@@ -1,12 +1,12 @@
-﻿@page "/NavMenu"
-@using FluentUI.Demo.Shared.Pages.NavMenu.Examples
+﻿@page "/NavMenuTree"
+@using FluentUI.Demo.Shared.Pages.NavMenuTree.Examples
 
-<h1>NavMenu, NavGroup and NavLink</h1>
+<h1>NavMenuTree, NavMenuGroup and NavMenuLink</h1>
 
 <div class="demopanel" style="text-align:center; margin-bottom: 10px">
     <b>IMPORTANT</b>
     <p>
-        With version 3.2 a new, much improved, set of components to build side-bar menus has been added.
+        With version 3.2 a new, much improved, set of components to build side-bar menus has been added. 
     </p>
     <p>
         <b>
@@ -15,7 +15,7 @@
         </b>
     </p>
     <p>
-        Everything works exactely as before by changing the name of this <code>FluentNavMenu</code> component in your application. (This probably needs to be done
+        Everything works exactely as before by changing the name of this <code>FluentNavMenu</code> component in your application. (This probably needs to be done 
         in one place only). The <code>FluentNavMenuGroup</code> and <code>FluentNavMenuLink</code> components have not been changed and do not need to be altered.
     </p>
     <p>
@@ -25,27 +25,26 @@
         If you wish to upgrade to the new menu components, please refer to the <a href="/WhatsNew">What's new</a> section for more information.
     </p>
 </div>
-
 <p>
-    The <code>FluentNavMenu</code>, <code>FluentNavGroup</code> and <code>FluentNavLink</code> components can be used to build
-    hierarchical, collapsible and expandable side-bar menus. They can range from simple 1-level deep lists to complex multi-level menu 
-    structures (with a max of 5 levels). 
+    The <code>FluentNavMenuGroup</code>, <code>FluentNavMenuLink</code> and <code>FluentNavMenu</code> components can be used to build
+    hierarchical, collapsible and expandable menus. They can range from simple 1-level deep lists to complex multi-level menu 
+    structures. 
 </p> 
 <p>
     None of these components are particulary useful when used stand-alone.
 </p>
 
-<ApiDocumentation Component="typeof(FluentNavMenu)" />
+<ApiDocumentation Component="typeof(FluentNavMenuTree)" />
 
-<ApiDocumentation Component="typeof(FluentNavGroup)" />
+<ApiDocumentation Component="typeof(FluentNavMenuGroup)" />
 
-<ApiDocumentation Component="typeof(FluentNavLink)" />
+<ApiDocumentation Component="typeof(FluentNavMenuLink)" />
 
 <h2>Examples</h2>
 
-<DemoSection Component="typeof(NavMenuDefault)" Title="Submenus, icons and plain">
+<DemoSection Component="typeof(NavMenuTreeDefault)" Title="Submenus, icons and plain">
     <Description>
-        This demo shows 3 different versions of a <code>NavMenu</code> (with <code>FluentNavGroup</code>s and <code>FluentNavLink</code>)s.
+        This demo shows 3 different versions of a <code>NavMenu</code> (with <code>FluentNavMenuGoups</code> and <code>FluentNavMenuLinks</code>).
         From left to right:
         <ul>
             <li>Menu with several sub-menus, icons, etc</li>
@@ -55,15 +54,20 @@
     </Description>
 </DemoSection>
 
-<DemoSection Component="typeof(NavMenuCollapsible)" Title="Collapsible navigation">
+<DemoSection Component="typeof(NavMenuTreeCollapsible)" Title="Collapsible navigation">
     <Description>
         More complete example which show how menu works with together with a text section when it collapses. 
     </Description>
 </DemoSection>
 
-<DemoSection Component="typeof(NavMenuDataBound)" Title="Data bound">
+<DemoSection Component="typeof(NavMenuTreeDataBound)" Title="Data bound">
     <Description>
         An example data binding the Expanded parameter.
     </Description>
 </DemoSection>
 
+<DemoSection Component="typeof(NavMenuTreeCustomActions)" Title="Custom actions">
+    <Description>
+        An example of intercepting menu actions to provide custom behavior.
+    </Description>
+</DemoSection>
