+++ examples/Demo/Shared/Shared/DemoMainLayout.razor
@@ -53,7 +53,7 @@
             <FluentIcon Icon="Icons.Regular.Size24.Navigation" />
         </label>
         <nav>
-            <DemoNavMenuTree />
+            <DemoNavMenu />
         </nav>
         <div class="content">
             <article id="article">
