+++ examples/Demo/Shared/Shared/DemoNavMenu.razor.css
@@ -1,0 +1,3 @@
+﻿::deep fluent-tree-item {
+    margin: 5px 0;
+}
