+++ examples/Demo/Shared/Shared/DemoNavMenuTree.razor
@@ -1,5 +1,5 @@
 <div>
-    <FluentNavMenu Collapsible="true" Title="Main menu" ReNavigate="true">
+    <FluentNavMenuTree Collapsible="true" Title="Main menu" ReNavigate="true">
         <FluentNavMenuLink Icon="@(new Icons.Regular.Size20.Home())" Href="/">
             <h2>Home</h2>
         </FluentNavMenuLink>
@@ -110,5 +110,5 @@
                 <FluentNavMenuLink Href="/Lab/TableOfContents" Icon="@(new Icons.Regular.Size20.DocumentTextLink())" Text="TableOfContents" />
             </FluentNavMenuGroup>
         </FluentNavMenuGroup>
-    </FluentNavMenu>
+    </FluentNavMenuTree>
 </div>
