+++ src/Core/Components/CollapsibleRegion/FluentCollapsibleRegion.razor
@@ -3,9 +3,5 @@
 @using System.Globalization
 
 <div @attributes="AdditionalAttributes" class="@ClassValue" style="@StyleValue">
-    <div @ref="_wrapper" class="fluent-collapsible-region-wrapper">
-        <div class="fluent-collapsible-region-wrapper-inner">
-            @ChildContent
-        </div>
-    </div>
+    @ChildContent
 </div>
