--- src/Core/Components/NavMenu2/FluentNavLink.razor
@@ -1,47 +1,0 @@
-﻿@namespace Microsoft.Fast.Components.FluentUI
-@inherits FluentComponentBase
-@using Microsoft.AspNetCore.Components.Routing
-
-<div @attributes="AdditionalAttributes" class="@ClassValue" disabled="@Disabled" style="@Style">
-    <div class="positioning-region">
-        <div class="content-region">
-            @if (!OnClick.HasDelegate)
-            {
-                <NavLink class="@LinkClassname"
-                         @attributes="@Attributes"
-                         Match="@Match"
-                         ActiveClass="@ActiveClass">
-                    @if (Icon is not null)
-                    {
-                        <FluentIcon Value="@Icon" Width="20px" Color="@IconColor" Class="fluent-nav-icon" />
-                    }
-                    else
-                    {
-                        <span class="fluent-nav-icon" style="min-width: 20px;"></span>
-                    }
-                    <div class="fluent-nav-text">
-                        @ChildContent
-                    </div>
-                </NavLink>
-            }
-            else
-            {
-                <div @onclick="OnClickHandler"
-                     class="@LinkClassname"
-                     ActiveClass="@ActiveClass">
-                    @if (Icon is not null)
-                    {
-                        <FluentIcon Value="@Icon" Width="20px" Color="@IconColor" />
-                    }
-else
-                    {
-                        <span class="fluent-nav-icon" style="min-width: 20px;"></span>
-                    }
-                    <div Class="fluent-nav-text">
-                        @ChildContent
-                    </div>
-                </div>
-            }
-        </div>
-    </div>
-</div>
