--- src/Core/Components/NavMenu2/FluentNavMenu2.razor
@@ -1,29 +1,0 @@
-﻿@namespace Microsoft.Fast.Components.FluentUI
-@inherits FluentComponentBase
-
-<div id="@Id" @attributes="AdditionalAttributes" class="@ClassValue" style="@StyleValue">
-    @if (Collapsible)
-    {
-        <div aria-label="@Title"
-             aria-expanded="@((!Collapsed).ToAttributeValue())"
-             class="fluent-nav-item"
-             @onclick="@ToggleCollapsedAsync"
-             @onkeydown="@HandleExpandCollapseKeyDownAsync">
-            <div class="positioning-region">
-                <div class="content-region">
-                    @if (CollapserContent is not null)
-                    {
-              
-                        @CollapserContent
-              
-                    }
-                    else
-                    {
-                        <FluentIcon Value="@(new CoreIcons.Regular.Size24.Navigation())" Width="20px" />
-                    }
-                </div>
-            </div>
-        </div>
-        }
-            @ChildContent
-    </div>
